+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.1.1
 - Bug fix: Preview text would trigger a disabled sound
 - Preview sound will trigger (text, fx and art)
+- New Feature: Level Up Trigger 
 
 # 0.1.0
 First release